pyproject.toml
[project]
name = "omni-webui"
dynamic = ["version"]
description = "An open-source web interface for LLM Chatbot."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiocache>=0.12.3",
    "aiofiles>=24.1.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "anyio[trio]>=4.7.0",
    "asyncpg>=0.30.0",
    "authlib>=1.4.0",
    "bcrypt>=4.2.1",
    "beautifulsoup4>=4.12.3",
    "black>=24.10.0",
    "boto3>=1.35.92",
    "chromadb>=0.6.1",
    "colbert-ai>=0.2.21",
    "duckduckgo-search>=7.2.0",
    "fastapi[all]>=0.115.6",
    "fpdf2>=2.8.2",
    "fsspec[s3]>=2024.12.0",
    "ftfy>=6.3.1",
    "jinja2>=3.1.5",
    "keepachangelog>=2.0.0",
    "langchain>=0.3.14",
    "langchain-community>=0.3.14",
    "ldap3>=2.9.1",
    "loguru>=0.7.3",
    "markdown>=3.7",
    "mcp>=1.2.0",
    "ollama>=0.4.5",
    "opensearch-py>=2.8.0",
    "pgvector>=0.3.6",
    "platformdirs>=4.3.6",
    "pydantic-extra-types[semver]>=2.10.1",
    "pydantic-settings>=2.7.1",
    "pydantic[email]>=2.10.4",
    "pydub>=0.25.1",
    "pyjwt>=2.10.1",
    "pymilvus>=2.5.3",
    "python-socketio>=5.12.1",
    "qdrant-client>=1.12.2",
    "redis>=5.2.1",
    "sentence-transformers>=3.3.1",
    "sqlalchemy[asyncio]>=2.0.36",
    "sqlmodel>=0.0.22",
    "swarmx>=0.4.1",
    "tiktoken>=0.8.0",
    "torch>=2.5.1",
    "typer>=0.15.1",
    "uvicorn[standard]>=0.34.0",
    "validators>=0.34.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Multimedia",
]

[project.scripts]
omni-webui = "omni_webui.__main__:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["omni_webui"]

[tool.hatch.version]
path = "package.json"
pattern = '"version":\s*"(?P<version>[^"]+)"'

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.5",
    "vulture>=2.14",
]
youtube-transcript-api = [
    "youtube-transcript-api>=0.6.3",
]
pdf = [
    "pypdf>=5.1.0",
]

[tool.pytest.ini_options]
addopts = [
    "--cov=omni_webui",
    "--cov-fail-under=90",
]
filterwarnings = "ignore:.*:DeprecationWarning\nignore:.*:PendingDeprecationWarning"

[tool.coverage.run]
omit = ["omni_webui/__main__.py", "omni_webui/app.py"]

[tool.coverage.report]
exclude_also = [
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "except importlib.metadata.PackageNotFoundError:\\n\\s+__version__ = \"0.0.0\"",
]

[tool.ruff.lint]
extend-select = ["I", "T20"]

[tool.pyright]
venvPath = "."
venv = ".venv"
